requirements.txt
pandas>=2.0.0
customtkinter>=5.2.2
Pillow>=10.0.0
playwright>=1.47.0
python-dotenv>=1.0.1
twocaptcha-python>=1.0.4
fdb>=2.0.2
openpyxl>=3.1.2
xlsxwriter>=3.2.0
